% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{calc_scaling_vector}
\alias{calc_scaling_vector}
\title{Calculate scaling factor for a dataset}
\usage{
calc_scaling_vector(data, scaling_factor, custom_scaling_vector, ncores)
}
\arguments{
\item{data}{dataset object}

\item{scaling_factor}{name of scaling factor; possible are: \code{census}, \code{spike-in} or \code{NONE} for no scaling factor}

\item{custom_scaling_vector}{named vector with custom scaling values for cell-types. Cell-types that do not occur in this vector but are present in the dataset will be set to 1}

\item{ncores}{number of cores}
}
\value{
a named vector with a scaling value for each cell in the dataset
}
\description{
Calculate scaling factor for a dataset
}
