% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_monocle}
\alias{census_monocle}
\title{Census calculation as implemented in monocle}
\usage{
census_monocle(expr_matrix, exp_capture_rate, expr_threshold)
}
\arguments{
\item{expr_matrix}{TPM matrix}

\item{exp_capture_rate}{expected capture rate; default=0.25}

\item{expr_threshold}{expression threshold; default=0}
}
\value{
vector with estimated mRNA values per cell in expr_matrix
}
\description{
Census calculation as implemented in monocle
}
