% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{check_annotation}
\alias{check_annotation}
\title{check for correct column names in annotation file and replace them if necessary}
\usage{
check_annotation(annotation, cell_column = "cell_type", id_column = 1)
}
\arguments{
\item{annotation}{dataframe; annotation dataframe}

\item{cell_column}{name of cell-type column; default is "cell_type"}

\item{id_column}{name of cell ID column; default is 1, which uses the rownames}
}
\value{
annotation dataframe with correct column names
}
\description{
check for correct column names in annotation file and replace them if necessary
}
