% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{check_if_tpm}
\alias{check_if_tpm}
\title{Checks, if a matrix is TPM-like (columns sum up to 1e6)}
\usage{
check_if_tpm(tpm_matrix, lower_limit = 7e+05)
}
\arguments{
\item{tpm_matrix}{matrix to check}

\item{lower_limit}{the lowest sum value, a cell may have}
}
\value{
boolean
}
\description{
Checks, if a matrix is TPM-like (columns sum up to 1e6)
}
\keyword{internal}
