% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{compare_matrix_with_annotation}
\alias{compare_matrix_with_annotation}
\title{Check if annotation and matrix have same cells}
\usage{
compare_matrix_with_annotation(m, annotation)
}
\arguments{
\item{m}{matrix, column names are cells}

\item{annotation}{data.frame, rownames are genes, cell names are in ID column}
}
\value{
intersected matrix
}
\description{
Otherwise intersection of both is used
}
