% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\docType{class}
\name{database-class}
\alias{database-class}
\title{class for a database (multiple datasets)}
\value{
a database object
}
\description{
class for a database (multiple datasets)
}
\section{Slots}{

\describe{
\item{\code{annotation}}{data.frame. a single dataframe with cell-type annotation of each cell in database}

\item{\code{counts}}{Matrix. a sparse matrix for the counts}

\item{\code{size}}{numeric. how many datasets are stored in the database object}
}}

