% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{database}
\alias{database}
\title{create a database class from multiple datasets}
\usage{
database(dataset_list)
}
\arguments{
\item{dataset_list}{list of datasets}
}
\value{
database object
}
\description{
create a database class from multiple datasets
}
