% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\title{class for a dataset}
\value{
a datasets object
}
\description{
This is the main class to store single-cell datasets and use them for pseudo-bulk simulations.
}
\section{Slots}{

\describe{
\item{\code{annotation}}{data.frame. a single dataframe with cell-type annotation of each cell in dataset}

\item{\code{counts}}{Matrix. a sparse matrix for the counts}

\item{\code{name}}{character. name of the dataset}
}}

