% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{basic constructor for a dataset}
\usage{
dataset(
  annotation,
  count_matrix,
  name,
  count_type = "raw",
  spike_in_col = NULL,
  read_number_col = NULL,
  additional_cols = NULL,
  filter_genes = T,
  variance_cutoff = 0,
  type_abundance_cutoff = 0
)
}
\arguments{
\item{annotation}{dataframe; needs columns 'ID' and 'cell_type'; 'ID' needs to be equal with cell_names in count_matrix}

\item{count_matrix}{sparse matrix; genes in rows, cells in columns}

\item{name}{name of the dataset; will be used for new unique IDs of cells}

\item{count_type}{what type of counts are in the dataset; default is 'TPM'}

\item{spike_in_col}{which column in annotation contains information on spike-in counts, which can be used to re-scale counts; mandatory for spike-in scaling factor in simulation}

\item{read_number_col}{which column in annotation contains information on total read numbers in a cell; mandatory for "spike-in" scaling factor and "read-number" in simulation}

\item{additional_cols}{list of column names in annotation, that should be stored as well in dataset object}

\item{filter_genes}{boolean, if TRUE, removes all genes with 0 expression over all samples & genes with variance below \code{variance_cutoff}}

\item{variance_cutoff}{numeric, is only applied if \code{filter_genes} is TRUE: removes all genes with variance below the chosen cutoff}

\item{type_abundance_cutoff}{numeric, remove all cells, whose cell-type appears less then the given value. This removes low abundant cell-types}
}
\value{
dataset object
}
\description{
basic constructor for a dataset
}
