% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_multiple}
\alias{dataset_multiple}
\title{constructor to merge multiple datasets into one}
\usage{
dataset_multiple(
  dataset_list,
  name,
  count_type = "raw",
  spike_in_col = NULL,
  read_number_col = NULL,
  additional_cols = NULL,
  filter_genes = T,
  variance_cutoff = 0,
  type_abundance_cutoff = 0
)
}
\arguments{
\item{dataset_list}{list of dataset objects}

\item{name}{name of the new dataset}
}
\value{
dataset object
}
\description{
constructor to merge multiple datasets into one
}
