% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_sfaira}
\alias{dataset_sfaira}
\title{constructor for a dataset using a sfaira ID}
\usage{
dataset_sfaira(
  sfaira_id,
  sfaira_setup,
  name,
  count_type = "TPM",
  spike_in_col = NULL,
  whitelist = NULL,
  annotation_column,
  id_column
)
}
\arguments{
\item{sfaira_id}{ID of a sfaira dataset}

\item{sfaira_setup}{the sfaira setup; given by \code{\link{setup_sfaira}}}

\item{name}{name of the dataset; will be used for new unique IDs of cells}

\item{count_type}{what type of counts are in the dataset; default is 'TPM'}

\item{spike_in_col}{which column in annotation contains information on spike-in counts, which can be used to re-scale counts}

\item{whitelist}{list of cell-types, which you want to use later for simulation; NULL to simulate all cell-types}

\item{annotation_column}{name of column in annotation with cell-types; should not be needed}

\item{id_column}{name of column in annotation with cell IDs; should not be needed}
}
\value{
dataset object
}
\description{
constructor for a dataset using a sfaira ID
}
