% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_sfaira_multiple}
\alias{dataset_sfaira_multiple}
\title{constructor for dataset using filters for tissue, organism and assay. This results in much larger datasets for simulation}
\usage{
dataset_sfaira_multiple(
  organisms = NULL,
  tissues = NULL,
  assays = NULL,
  sfaira_setup,
  name,
  count_type = "TPM",
  spike_in_col = NULL
)
}
\arguments{
\item{organisms}{list of organisms (only human and mouse available)}

\item{tissues}{list of tissues}

\item{assays}{list of assays}

\item{sfaira_setup}{the sfaira setup; given by \code{\link{setup_sfaira}}}

\item{name}{name of the dataset; will be used for new unique IDs of cells}

\item{count_type}{what type of counts are in the dataset; default is 'TPM'}

\item{spike_in_col}{which column in annotation contains information on spike-in counts, which can be used to re-scale counts}
}
\value{
dataset object
}
\description{
constructor for dataset using filters for tissue, organism and assay. This results in much larger datasets for simulation
}
