% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{dmode}
\alias{dmode}
\title{use gaussian kernel to calculate the mode of transcript counts}
\usage{
dmode(x)
}
\arguments{
\item{x}{vector of numeric values}
}
\value{
most commonly occurring (log-transformed) TPM value
}
\description{
use gaussian kernel to calculate the mode of transcript counts
}
\keyword{internal}
