% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sfaira.R
\name{download_sfaira}
\alias{download_sfaira}
\title{download a specific dataset from sfaira by an ID}
\usage{
download_sfaira(
  setup_list,
  ids,
  force = FALSE,
  synapse_user = NULL,
  synapse_pw = NULL
)
}
\arguments{
\item{setup_list}{the sfaira setup; given by \code{\link{setup_sfaira}}}

\item{ids}{the IDs of the datasets}

\item{force}{logical; TRUE if you want to force the download, even though no cell-type annotation exists for this dataset. Default if FALSE}

\item{synapse_user}{character; username for synapse portal (https://www.synapse.org)}

\item{synapse_pw}{character; password for synapse portal (https://www.synapse.org)}
}
\value{
matrix, gene names and cell IDs
}
\description{
download a specific dataset from sfaira by an ID
}
