% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sfaira.R
\name{download_sfaira_multiple}
\alias{download_sfaira_multiple}
\title{download multiple datasets from sfaira using filters for organism, tissue and/or assay}
\usage{
download_sfaira_multiple(
  setup_list,
  organisms = NULL,
  tissues = NULL,
  assays = NULL,
  force = F
)
}
\arguments{
\item{setup_list}{the sfaira setup; given by \code{\link{setup_sfaira}}}

\item{organisms}{list of organisms (only human and mouse available)}

\item{tissues}{list of tissues}

\item{assays}{list of assays}

\item{force}{logical; TRUE if you want to force to download all datasets, otherwise only the ones with cell-type annotation will be returned. Default if FALSE}
}
\value{
annotated data object, contains count matrix and annotation
}
\description{
similar to the filters on the sfaira website (\url{https://theislab.github.io/sfaira-portal/Datasets})
}
