% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{filter_matrix}
\alias{filter_matrix}
\title{filter one (or two) expression matrix by genes}
\usage{
filter_matrix(m1, m2 = NULL, filter_genes = TRUE, variance_cutoff = 0)
}
\arguments{
\item{m1}{Matrix 1}

\item{m2}{Matrix 2 (optional)}

\item{filter_genes}{boolean}

\item{variance_cutoff}{numeric, genes below this variance value are removed}
}
\value{
filtered matrix
}
\description{
filter one (or two) expression matrix by genes
}
\keyword{internal}
