% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{generate_summarized_experiment}
\alias{generate_summarized_experiment}
\title{Generate SummarizedExperiment using multiple parameters}
\usage{
generate_summarized_experiment(
  annotation,
  count_matrix,
  tpm_matrix,
  name,
  spike_in_col,
  read_number_col,
  additional_cols,
  filter_genes,
  variance_cutoff,
  type_abundance_cutoff,
  scale_tpm
)
}
\value{
Return a \link[SummarizedExperiment]{SummarizedExperiment} object
}
\description{
Generate SummarizedExperiment using multiple parameters
}
