% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{h5ad_to_adata}
\alias{h5ad_to_adata}
\title{Use basilisk environment to read h5ad file and access anndata object}
\usage{
h5ad_to_adata(h5ad_path)
}
\arguments{
\item{h5ad_path}{path to h5ad file}
}
\value{
matrix contained on h5ad file as dgCMatrix
}
\description{
Use basilisk environment to read h5ad file and access anndata object
}
