% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{merge_scaling_factor}
\alias{merge_scaling_factor}
\title{Create scaling vector from custom or pre-defined scaling factor}
\usage{
merge_scaling_factor(data, scaling_factor_values, scaling_factor_name)
}
\arguments{
\item{data}{dataset}

\item{scaling_factor_values}{named list of scaling values}

\item{scaling_factor_name}{name of scaling factor method}
}
\value{
scaling vector
}
\description{
Create scaling vector from custom or pre-defined scaling factor
}
