% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{merge_simulations}
\alias{merge_simulations}
\title{Combine multiple simulations into one result}
\usage{
merge_simulations(simulation_list)
}
\arguments{
\item{simulation_list}{a list of simulations}
}
\value{
named list; \code{bulk} a \link[SummarizedExperiment]{SummarizedExperiment} object, where the assays store the simulated bulk RNAseq datasets. Can hold either one or two assays, depending on how many matrices were present in the dataset
\code{cell-fractions} is a dataframe with the simulated cell-fractions per sample;
\code{scaling_vector} scaling value for each cell in dataset
}
\description{
we recommend to only merge simulations from the same dataset object, otherwise the count matrices might not correspond on the gene level
}
