% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{plot_simulation}
\alias{plot_simulation}
\title{Plot the cell-type fractions in your simulated dataset}
\usage{
plot_simulation(simulation)
}
\arguments{
\item{simulation}{a simulation object generated by \code{simulate_bulk}}
}
\value{
a gpplot2 barplot
}
\description{
Plot the cell-type fractions in your simulated dataset
}
