% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{sample_cells_by_read_depth}
\alias{sample_cells_by_read_depth}
\title{Sample cells based of sequencing depth}
\usage{
sample_cells_by_read_depth(cells, depth)
}
\arguments{
\item{cells}{dataframe; needs to contain a "read_number" column}

\item{total_read_counts}{numeric; the total number of reads the sampled cells will sum up to (approximatly)}
}
\value{
list of cell IDs
}
\description{
This function allows to sample cells from a dataframe including the number of reads per cell until a pre-defined
sequencing depth is met or no cells are left with low enough read numbers to reach this depth.
}
