% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{save_simulation}
\alias{save_simulation}
\title{Save the expression matrix of a simulated pseudo-bulk dataset to a file}
\usage{
save_simulation(simulation, filename)
}
\arguments{
\item{simulation}{the result of simulate_bulk()}

\item{filename}{the filename where to save the expression matrix to}
}
\description{
Save the expression matrix of a simulated pseudo-bulk dataset to a file
}
