% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sfaira.R
\name{setup_sfaira}
\alias{setup_sfaira}
\title{setup the sfaira package}
\usage{
setup_sfaira(python_path, env_name, basedir)
}
\arguments{
\item{python_path}{path to the python executable}

\item{env_name}{name of the conda environment in which sfaira is installed}

\item{basedir}{name of the directory, where the raw files will be downloaded into}
}
\value{
list with sfaira package and file directories; must be used for other sfaira functions
}
\description{
setup the sfaira package
}
