% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sfaira.R
\name{sfaira_overview}
\alias{sfaira_overview}
\title{Gives an overview of the possible datasets you can use from the sfaira database}
\usage{
sfaira_overview(setup_list)
}
\arguments{
\item{setup_list}{the sfaira setup; given by \code{\link{setup_sfaira}}}
}
\value{
a dataframe with information on each dataset
}
\description{
Gives an overview of the possible datasets you can use from the sfaira database
}
\examples{
setup_list <- setup_sfaira(basedir = tempdir())
all_datasets <- sfaira_overview(setup_list)
}
