% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census}
\alias{census}
\title{applies the census count transformation on a count matrix}
\usage{
census(
  matrix,
  exp_capture_rate = 0.25,
  expr_threshold = 0,
  ncores = 1,
  method = c("monocle", "paper")
)
}
\arguments{
\item{matrix}{sparse count matrix; cells in columns, genes in rows}

\item{exp_capture_rate}{expected capture rate; default=0.25}

\item{expr_threshold}{expression threshold; default=0}

\item{ncores}{number of cores}

\item{method}{implementation of census; possible are: \code{monocle} and \code{paper}}
}
\value{
a vector for each cell-type, with a scaling factor which can be used to transform the counts of the matrix
}
\description{
needs a sparse matrix with cells in columns and genes in rows. You can find the detailed explaination here:
\url{http://cole-trapnell-lab.github.io/monocle-release/docs/#census}
}
