% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_h5ad}
\alias{dataset_h5ad}
\title{constructor for a dataset using a h5ad file for the counts}
\usage{
dataset_h5ad(
  annotation,
  h5ad_file,
  name,
  count_type = "TPM",
  spike_in_col = NULL
)
}
\arguments{
\item{annotation}{dataframe; needs columns 'ID' and 'cell_type'; 'ID' needs to be equal with cell_names in count_matrix}

\item{h5ad_file}{h5ad file with count data}

\item{name}{name of the dataset; will be used for new unique IDs of cells}

\item{count_type}{what type of counts are in the dataset; default is 'TPM'}

\item{spike_in_col}{which column in annotation contains information on spike-in counts, which can be used to re-scale counts}
}
\value{
dataset object
}
\description{
constructor for a dataset using a h5ad file for the counts
}
