% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_h5ad}
\alias{dataset_h5ad}
\title{constructor for a dataset using a h5ad file for the counts}
\usage{
dataset_h5ad(
  annotation,
  h5ad_file,
  name,
  count_type = "raw",
  spike_in_col = NULL,
  filter_genes = T,
  variance_cutoff = 0
)
}
\arguments{
\item{annotation}{dataframe; needs columns 'ID' and 'cell_type'; 'ID' needs to be equal with cell_names in count_matrix}

\item{h5ad_file}{h5ad file with count data}

\item{name}{name of the dataset; will be used for new unique IDs of cells}

\item{count_type}{what type of counts are in the dataset; default is 'TPM'}

\item{spike_in_col}{which column in annotation contains information on spike-in counts, which can be used to re-scale counts}

\item{filter_genes}{boolean, if TRUE, removes all genes with 0 expression over all samples & genes with variance below \code{variance_cutoff}}

\item{variance_cutoff}{numeric, is only applied if \code{filter_genes} is TRUE: removes all genes with variance below the chosen cutoff}
}
\value{
dataset object
}
\description{
constructor for a dataset using a h5ad file for the counts
}
