% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_merge}
\alias{dataset_merge}
\title{Merge multiple \link[SummarizedExperiment]{SummarizedExperiment} datasets into one}
\usage{
dataset_merge(
  dataset_list,
  name = "SimBu_dataset",
  spike_in_col = NULL,
  read_number_col = NULL,
  additional_cols = NULL,
  filter_genes = T,
  variance_cutoff = 0,
  type_abundance_cutoff = 0,
  scale_tpm = T
)
}
\arguments{
\item{dataset_list}{list of \link[SummarizedExperiment]{SummarizedExperiment} objects}

\item{name}{name of the new dataset}
}
\value{
\link[SummarizedExperiment]{SummarizedExperiment} object
}
\description{
Merge multiple \link[SummarizedExperiment]{SummarizedExperiment} datasets into one
}
