% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_seurat}
\alias{dataset_seurat}
\title{constructor for dataset using a seurat object for the counts}
\usage{
dataset_seurat(
  annotation,
  seurat_obj,
  name,
  count_type = "raw",
  spike_in_col = NULL,
  read_number_col = NULL,
  additional_cols = NULL,
  filter_genes = T,
  variance_cutoff = 0,
  type_abundance_cutoff = 0
)
}
\arguments{
\item{annotation}{dataframe; needs columns 'ID' and 'cell_type'; 'ID' needs to be equal with cell_names in count_matrix}

\item{seurat_obj}{Seurat object; needs to have counts at position assays$RNA$counts}

\item{name}{name of the dataset; will be used for new unique IDs of cells}

\item{count_type}{what type of counts are in the dataset; default is 'TPM'}

\item{spike_in_col}{which column in annotation contains information on spike-in counts, which can be used to re-scale counts}

\item{read_number_col}{which column in annotation contains information on total read numbers in a cell; mandatory for "spike-in" scaling factor and "read-number" in simulation}

\item{filter_genes}{boolean, if TRUE, removes all genes with 0 expression over all samples & genes with variance below \code{variance_cutoff}}

\item{variance_cutoff}{numeric, is only applied if \code{filter_genes} is TRUE: removes all genes with variance below the chosen cutoff}

\item{type_abundance_cutoff}{numeric, remove all cells, whose cell-type appears less then the given value. This removes low abundant cell-types}
}
\value{
dataset object
}
\description{
constructor for dataset using a seurat object for the counts
}
