% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_seurat}
\alias{dataset_seurat}
\title{Function to generate a \link[SummarizedExperiment]{SummarizedExperiment} using a \link[Seurat]{Seurat} object}
\usage{
dataset_seurat(
  seurat_obj,
  count_assay,
  cell_id_col,
  cell_type_col,
  tpm_assay = NULL,
  name = "SimBu_dataset",
  spike_in_col = NULL,
  additional_cols = NULL,
  filter_genes = T,
  variance_cutoff = 0,
  type_abundance_cutoff = 0,
  scale_tpm = T
)
}
\arguments{
\item{seurat_obj}{(mandatory) \link[Seurat]{Seurat} object with TPM counts}

\item{count_assay}{(mandatory) name of assay in Seurat object which contains count data in 'counts' slot}

\item{cell_id_col}{(mandatory) name of column in Seurat meta.data with unique cell ids}

\item{cell_type_col}{(mandatory) name of column in Seurat meta.data with cell type name}

\item{tpm_assay}{name of assay in Seurat object which contains TPM data in 'counts' slot}

\item{name}{name of the dataset; will be used for new unique IDs of cells}

\item{spike_in_col}{which column in annotation contains information on spike_in counts, which can be used to re-scale counts; mandatory for spike_in scaling factor in simulation}

\item{additional_cols}{list of column names in annotation, that should be stored as well in dataset object}

\item{filter_genes}{boolean, if TRUE, removes all genes with 0 expression over all samples & genes with variance below \code{variance_cutoff}}

\item{variance_cutoff}{numeric, is only applied if \code{filter_genes} is TRUE: removes all genes with variance below the chosen cutoff}

\item{type_abundance_cutoff}{numeric, remove all cells, whose cell-type appears less then the given value. This removes low abundant cell-types}

\item{scale_tpm}{boolean, if TRUE (default) the cells in tpm_matrix will be scaled to sum up to 1e6}
}
\value{
Return a \link[SummarizedExperiment]{SummarizedExperiment} object
}
\description{
Function to generate a \link[SummarizedExperiment]{SummarizedExperiment} using a \link[Seurat]{Seurat} object
}
