% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_sfaira}
\alias{dataset_sfaira}
\title{constructor for a dataset using a sfaira ID}
\usage{
dataset_sfaira(
  sfaira_id,
  sfaira_setup,
  name,
  count_type = "TPM",
  spike_in_col = NULL,
  force = F,
  filter_genes = F
)
}
\arguments{
\item{sfaira_id}{ID of a sfaira dataset}

\item{sfaira_setup}{the sfaira setup; given by \code{\link{setup_sfaira}}}

\item{name}{name of the dataset; will be used for new unique IDs of cells}

\item{count_type}{what type of counts are in the dataset; default is 'TPM'}

\item{spike_in_col}{which column in annotation contains information on spike-in counts, which can be used to re-scale counts}

\item{filter_genes}{boolean, if zero expression genes will be included in dataset or not}
}
\value{
dataset object
}
\description{
constructor for a dataset using a sfaira ID
}
