% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_sfaira}
\alias{dataset_sfaira}
\title{Build a dataset using a single sfaira entry ID}
\usage{
dataset_sfaira(
  sfaira_id,
  sfaira_setup,
  name,
  count_type = "raw",
  spike_in_col = NULL,
  read_number_col = NULL,
  additional_cols = NULL,
  force = F,
  filter_genes = T,
  variance_cutoff = 0,
  type_abundance_cutoff = 0
)
}
\arguments{
\item{sfaira_id}{ID of a sfaira dataset}

\item{sfaira_setup}{the sfaira setup; given by \code{\link{setup_sfaira}}}

\item{name}{name of the dataset; will be used for new unique IDs of cells}

\item{count_type}{what type of counts are in the dataset; default is 'TPM'}

\item{spike_in_col}{which column in annotation contains information on spike-in counts, which can be used to re-scale counts}

\item{read_number_col}{which column in annotation contains information on total read numbers in a cell; mandatory for "spike-in" scaling factor and "read-number" in simulation}

\item{filter_genes}{boolean, if TRUE, removes all genes with 0 expression over all samples & genes with variance below \code{variance_cutoff}}

\item{variance_cutoff}{numeric, is only applied if \code{filter_genes} is TRUE: removes all genes with variance below the chosen cutoff}

\item{type_abundance_cutoff}{numeric, remove all cells, whose cell-type appears less then the given value. This removes low abundant cell-types}
}
\value{
dataset object
}
\description{
Build a dataset using a single sfaira entry ID
}
