% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sfaira.R
\name{download_sfaira}
\alias{download_sfaira}
\title{download a specific dataset from sfaira by an ID}
\usage{
download_sfaira(setup_list, id, force = F)
}
\arguments{
\item{setup_list}{the sfaira setup; given by \code{\link{setup_sfaira}}}

\item{id}{the ID of the datasets}

\item{force}{logical; TRUE if you want to force the download, even though no cell-type annotation exists for this dataset. Default if FALSE}
}
\value{
a anndata object, stores counts, metadata and other information on the dataset
}
\description{
download a specific dataset from sfaira by an ID
}
