% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{merge_simulations}
\alias{merge_simulations}
\title{Combine multiple simulations into one result}
\usage{
merge_simulations(simulation_list)
}
\arguments{
\item{simulation_list}{a list of simulations}
}
\value{
named list; \code{pseudo_bulk} is a sparse matrix with the simulated counts;
\code{cell-fractions} is a dataframe with the simulated cell-fractions per sample;
\code{expression_set} is a Bioconductor Expression Set \url{http://www.bioconductor.org/packages/release/bioc/vignettes/Biobase/inst/doc/ExpressionSetIntroduction.pdf}
}
\description{
we recommend to only merge simulations from the same dataset object, otherwise the count matrices might not correspond on the gene level
}
