% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sfaira.R
\name{setup_sfaira}
\alias{setup_sfaira}
\title{setup the sfaira package}
\usage{
setup_sfaira(basedir)
}
\arguments{
\item{basedir}{name of the directory, where the raw files will be downloaded into}
}
\value{
list with sfaira package and file directories; must be used for other sfaira functions
}
\description{
If you want to download datasets from Sfaira, you need to specify a directory where the datasets are saved into.
Additionally, when this function is called for the first time, a conda environment will be established and sfaira along
all of its dependencies are installed. This can take some time but will be only performed one single time, as the
environment can be re-used.
}
\examples{
\dontrun{
sfaira_list <- setup_sfaira(basedir=tempdir())
}
}
