% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_bulk}
\alias{simulate_bulk}
\title{simulate whole pseudo-bulk RNAseq dataset}
\usage{
simulate_bulk(
  data,
  scenario = c("noisy", "random", "spike-in", "spill-over"),
  scaling_factor = c("NONE", "census", "spike-in", "custom"),
  spike_in_cell_type = NULL,
  spike_in_amount = NULL,
  spillover_cell_type = NULL,
  nsamples = 100,
  ncells = 1000,
  ncores = 1
)
}
\arguments{
\item{data}{\code{\link{database}} or \code{\link{dataset}} object}

\item{scenario}{select on of the pre-defined cell-type fraction scenarios; possible are: \code{noisy},\code{random},\code{spike-in},\code{spill-over}}

\item{scaling_factor}{name of scaling factor; possible are: \code{census}, \code{spike-in}, \code{custom}}

\item{spike_in_cell_type}{name of cell-type used for \code{spike-in} scenario}

\item{spike_in_amount}{fraction of cell-type used for \code{spike-in} scenario; must be between \code{0} and \code{0.99}}

\item{spillover_cell_type}{name of cell-type used for \code{spill-over} scenario}

\item{nsamples}{numeric; number of samples in pseudo-bulk RNAseq dataset}

\item{ncells}{numeric; number of cells in each dataset}

\item{ncores}{numeric; number of cores to use}
}
\value{
named list; \code{pseudo_bulk} is a sparse matrix with the simulated counts;
\code{cell-fractions} is a dataframe with the simulated cell-fractions per sample;
\code{expression_set} is a Bioconductor Expression Set \url{http://www.bioconductor.org/packages/release/bioc/vignettes/Biobase/inst/doc/ExpressionSetIntroduction.pdf}
}
\description{
simulate whole pseudo-bulk RNAseq dataset
}
