% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_sample}
\alias{simulate_sample}
\title{function to sample cells according to given cell-type fractions
Note: if total_read_counts is used, the cell-fractions are applied to the number of counts, not the number of cells!}
\usage{
simulate_sample(
  data,
  scaling_factor,
  simulation_vector,
  total_cells,
  total_read_counts,
  ncores
)
}
\arguments{
\item{data}{\code{\link{database}} or \code{\link{dataset}} object}

\item{scaling_factor}{name of scaling factor; possible are: \code{census}, \code{spike-in}, \code{custom}}

\item{simulation_vector}{named vector with wanted cell-types and their fractions}

\item{total_cells}{numeric; number of total cells for this simulation}

\item{total_read_counts}{numeric; sets the total read count value for each sample}

\item{ncores}{numeric; number of cores used to create simulation}
}
\value{
returns a vector with mean count/TPM values over all selected samples per gene
}
\description{
function to sample cells according to given cell-type fractions
Note: if total_read_counts is used, the cell-fractions are applied to the number of counts, not the number of cells!
}
