% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_sample}
\alias{simulate_sample}
\title{simulate single pseudo-bulk sample}
\usage{
simulate_sample(
  data,
  scaling_factor,
  scaling_vector,
  simulation_vector,
  scaling_factor_aggregation,
  total_cells,
  total_read_counts,
  ncores
)
}
\arguments{
\item{data}{\code{\link{database}} or \code{\link{dataset}} object}

\item{scaling_factor}{name of scaling factor; possible are: \code{census}, \code{spike-in}, \code{custom}}

\item{scaling_vector}{vector with scaling values for each cell; calculated by the \code{calc_scaling_vector} function}

\item{simulation_vector}{named vector with wanted cell-types and their fractions}

\item{scaling_factor_aggregation}{aggregation function how to apply the scaling factor, possible are: \code{multiply}(default) and \code{division}}

\item{total_cells}{numeric; number of total cells for this simulation}

\item{total_read_counts}{numeric; sets the total read count value for each sample}

\item{ncores}{numeric; number of cores used to create simulation}

\item{sample_aggreation}{aggregation method on how to generate a sample; possible are: \code{mean}(default), \code{sum}, \code{median}}
}
\value{
returns a vector with expression values for all genes in the provided dataset
}
\description{
function to sample cells according to given cell-type fractions. This creates a single pseudo-bulk sample by calculating the
mean expression value per gene over all sampled cells.
Note: if total_read_counts is used, the cell-fractions are applied to the number of counts, not the number of cells!
}
